PicktureDeleteBackground_bot/requirements.txt
aiogram==3.7.0
rembg
pillow
onnxruntime